refactor(tracking): move fns from fetcher to project helper
