+++ packages/orbit-tracking/src/helpers/index.ts
@@ -1,19 +1,19 @@
 import { errorMessage, infoMessage } from "./message";
 import { getCategory } from "./category";
-import { filterMembers } from "./members";
+import { mapProjects } from "./projects";
+import { projectCmd } from "./project";
 import { headingTemplate } from "./template";
 import { __dirname, getOutputPath } from "./path";
 import { timestamp } from "./timestamp";
-import { getVersions } from "./versions";
 
 export {
   getCategory,
-  filterMembers,
   errorMessage,
   infoMessage,
+  mapProjects,
+  projectCmd,
   headingTemplate,
   getOutputPath,
   __dirname,
   timestamp,
-  getVersions,
 };
